backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
sqlalchemy
pandas
openpyxl
python-multipart
scikit-learn